+++ frontend/lib/resolution/Makefile.include
@@ -21,8 +21,8 @@ FRONTEND_RESOLUTION_OBJDIR = $(COMPILER_BUILD)/frontend/lib/resolution
 ALL_SRCS += ../frontend/lib/resolution/*.h ../frontend/lib/resolution/*.cpp
 
 FRONTEND_RESOLUTION_SRCS =                          \
-           CallInitDeinit.cpp                       \
            Resolver.cpp                             \
+           call-init-deinit.cpp                     \
            can-pass.cpp                             \
            default-functions.cpp                    \
            disambiguation.cpp                       \
