+++ frontend/lib/resolution/resolution-queries.cpp
@@ -31,7 +31,7 @@
 #include "chpl/types/all-types.h"
 #include "chpl/uast/all-uast.h"
 
-#include "CallInitDeinit.h"
+#include "call-init-deinit.h"
 #include "Resolver.h"
 #include "default-functions.h"
 #include "prims.h"
@@ -1695,7 +1695,7 @@ resolveFunctionByInfoQuery(Context* context,
     auto newTfsForInitializer = visitor.initResolver->finalize();
 
     // then, resolve '=' and add any copy init/deinit calls as needed
-    CallInitDeinit::process(visitor);
+    callInitDeinit(visitor);
 
     // TODO: can this be encapsulated in a method?
     resolvedPoiInfo.swap(visitor.poiInfo);
